refactor test files
