--- server/test/mocha.opts
@@ -1,4 +1,0 @@
--r @babel/polyfill
--r @babel/register
---recursive
---exit
