+++ server/test/routes.user.spec.js
@@ -1,6 +1,6 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import app from '../../src/app';
+import app from '../src/app';
 
 chai.use(chaiHttp);
